modified-Reservation計算住宿費用，更改Reservation、RoomType資料表結構

+++ Infrastructure/Models/RoomType.cs
@@ -18,8 +18,8 @@ namespace Infrastructure.Models
         public string Name { get; set; }
         public double Area { get; set; }
         public int BedQuantity { get; set; }
-        public decimal Price { get; set; }
-        public decimal Hprice { get; set; }
+        public int Price { get; set; }
+        public int Hprice { get; set; }
         public bool AirCondition { get; set; }
         public bool Tv { get; set; }
         public DateTime SysDate { get; set; }
